Check for output folder before clearing
In normal operation there should always be an output folder for
clear_output_folder() to delete, but in case there isn't (maybe an error
in testing, or it is deleted by the user) the program will no longer
crash upon trying to empty a folder which doesn't exist.

+++ library/paintjob.py
@@ -31,7 +31,8 @@ class Pack:
 
 def clear_output_folder():
     print("Clearing output folder")
-    shutil.rmtree("output")
+    if os.path.exists("output"):
+        shutil.rmtree("output")
     os.makedirs("output")
 
 def make_folder(path):
